requirements.txt
Flask
Flask-Cors
Pillow
torch
transformers
torchvision
numpy
scikit-learn
pandas
joblib
